backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-groq
ddgs
tqdm
gunicorn
# --- RAG dependencies ---
pypdf
sentence-transformers
langchain-huggingface
faiss-cpu
python-docx
docx2txt